Moved net methods and calls around to optimize access from other programs when used as a library. Made adjustments for MessageContext. Redid /ban, /kick, /warn, /unban. Added CommandContext for easier argument parsing.

+++ UncreatedWarfare/Stats/StatsCoroutine.cs
@@ -74,7 +74,7 @@ namespace Uncreated.Warfare.Stats
                                 }
                                 else if (afk.time == Afk.Clamp(n + 1)) // one cycle left
                                 {
-                                    ucplayer.SendChat("afk_warning", ((uint)UCWarfare.Config.StatsInterval).GetTimeFromMinutes(ucplayer.Steam64));
+                                    ucplayer.SendChat("afk_warning", UCWarfare.Config.StatsInterval.GetTimeFromMinutes(ucplayer.Steam64));
                                 }
                             }
                             else
